Move AnnotationsPage into a module
