+++ metaspace/webapp/src/modules/Annotations/ProgressButton.vue
