+++ metaspace/webapp/src/modules/Annotations/annotation-widgets/IsotopePatternPlot.vue
