+++ metaspace/webapp/src/modules/Annotations/annotation-widgets/PlotLegend.vue
