+++ metaspace/webapp/src/modules/Annotations/annotation-widgets/default/IonImageSettings.vue
