+++ metaspace/webapp/src/modules/Annotations/annotation-widgets/lcms/MainImageHeader.vue
