+++ metaspace/webapp/src/modules/Annotations/annotation-widgets/lcms/XicPlot.vue
